// client/src/Gen/models/MealMacronutrientEntryPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nutrition Tracker API
 * API Specification for Nutrition Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface MealMacronutrientEntryPostRequest
 */
export interface MealMacronutrientEntryPostRequest {
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryPostRequest
     */
    protein: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryPostRequest
     */
    carbohydrate: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryPostRequest
     */
    fat: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryPostRequest
     */
    kCal?: number;
}

/**
 * Check if a given object implements the MealMacronutrientEntryPostRequest interface.
 */
export function instanceOfMealMacronutrientEntryPostRequest(value: object): value is MealMacronutrientEntryPostRequest {
    if (!("protein" in value) || value["protein"] === undefined) return false;
    if (!("carbohydrate" in value) || value["carbohydrate"] === undefined) return false;
    if (!("fat" in value) || value["fat"] === undefined) return false;
    return true;
}

export function MealMacronutrientEntryPostRequestFromJSON(json: any): MealMacronutrientEntryPostRequest {
    return MealMacronutrientEntryPostRequestFromJSONTyped(json, false);
}

export function MealMacronutrientEntryPostRequestFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): MealMacronutrientEntryPostRequest {
    if (json == null) {
        return json;
    }
    return {
        protein: json["protein"],
        carbohydrate: json["carbohydrate"],
        fat: json["fat"],
        kCal: json["kCal"] == null ? undefined : json["kCal"],
    };
}

export function MealMacronutrientEntryPostRequestToJSON(json: any): MealMacronutrientEntryPostRequest {
    return MealMacronutrientEntryPostRequestToJSONTyped(json, false);
}

export function MealMacronutrientEntryPostRequestToJSONTyped(
    value?: MealMacronutrientEntryPostRequest | null,
    ignoreDiscriminator: boolean = false,
): any {
    if (value == null) {
        return value;
    }

    return {
        protein: value["protein"],
        carbohydrate: value["carbohydrate"],
        fat: value["fat"],
        kCal: value["kCal"],
    };
}