// client/src/Gen/models/MealMacronutrientEntryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nutrition Tracker API
 * API Specification for Nutrition Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface MealMacronutrientEntryResponse
 */
export interface MealMacronutrientEntryResponse {
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryResponse
     */
    id: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryResponse
     */
    protein: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryResponse
     */
    carbohydrate: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryResponse
     */
    fat: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryResponse
     */
    kCal: number;
    /**
     *
     * @type {number}
     * @memberof MealMacronutrientEntryResponse
     */
    sequenceNumber: number;
}

/**
 * Check if a given object implements the MealMacronutrientEntryResponse interface.
 */
export function instanceOfMealMacronutrientEntryResponse(value: object): value is MealMacronutrientEntryResponse {
    if (!("id" in value) || value["id"] === undefined) return false;
    if (!("protein" in value) || value["protein"] === undefined) return false;
    if (!("carbohydrate" in value) || value["carbohydrate"] === undefined) return false;
    if (!("fat" in value) || value["fat"] === undefined) return false;
    if (!("kCal" in value) || value["kCal"] === undefined) return false;
    if (!("sequenceNumber" in value) || value["sequenceNumber"] === undefined) return false;
    return true;
}

export function MealMacronutrientEntryResponseFromJSON(json: any): MealMacronutrientEntryResponse {
    return MealMacronutrientEntryResponseFromJSONTyped(json, false);
}

export function MealMacronutrientEntryResponseFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): MealMacronutrientEntryResponse {
    if (json == null) {
        return json;
    }
    return {
        id: json["id"],
        protein: json["protein"],
        carbohydrate: json["carbohydrate"],
        fat: json["fat"],
        kCal: json["kCal"],
        sequenceNumber: json["sequenceNumber"],
    };
}

export function MealMacronutrientEntryResponseToJSON(json: any): MealMacronutrientEntryResponse {
    return MealMacronutrientEntryResponseToJSONTyped(json, false);
}

export function MealMacronutrientEntryResponseToJSONTyped(
    value?: MealMacronutrientEntryResponse | null,
    ignoreDiscriminator: boolean = false,
): any {
    if (value == null) {
        return value;
    }

    return {
        id: value["id"],
        protein: value["protein"],
        carbohydrate: value["carbohydrate"],
        fat: value["fat"],
        kCal: value["kCal"],
        sequenceNumber: value["sequenceNumber"],
    };
}